requirements.txt
pandas>=2.0.0
pillow>=9.0.0
easyocr>=1.7.0
pdf2image>=1.16.0
openai>=1.0.0
numpy>=2.0.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
matplotlib>=3.5.0
# HAL 9000 Terminal Interface - No additional dependencies required
# Uses only Python standard library modules: time, random, os